refactor(main): extract per-file processors into processors/ (step 2)
Step 2 of #127. Move per-file work out of src/main.rs into focused
processor modules:

  src/processors/utils.rs      apply_with_temp_file, restore_timestamp
  src/processors/apply.rs      process_apply, process_apply_channel
  src/processors/info.rs       process_info
  src/processors/undo.rs       process_undo
  src/processors/replaygain.rs process_track_gain,
                               process_apply_replaygain_with_album,
                               process_apply_replaygain_aac_with_album

main.rs shrinks from 2307 → 1257 lines. The cmd_* dispatchers stay
in main.rs and will move in step 3.

get_filename is promoted to pub(crate) so processors/* can share
the same path-to-display-name helper as the still-resident cmd_*
functions.

No public library API change. CLI behavior unchanged.
