+++ src/main.rs
@@ -5,37 +5,39 @@
 
 mod cli;
 mod json_output;
+mod processors;
 mod progress;
 
 use anyhow::Result;
 use colored::*;
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
-use mp3rgain::aac;
 use mp3rgain::id3v2;
 use mp3rgain::mp4meta;
-use mp3rgain::replaygain::{self, AudioFileType, ReplayGainResult, REPLAYGAIN_REFERENCE_DB};
+use mp3rgain::replaygain::{self, REPLAYGAIN_REFERENCE_DB};
 use mp3rgain::{
     analyze, db_to_steps, delete_ape_tag, find_max_amplitude, read_ape_tag_from_file, steps_to_db,
-    undo_gain, Channel, GainOptions, TAG_MP3GAIN_MINMAX, TAG_MP3GAIN_UNDO,
-    TAG_REPLAYGAIN_ALBUM_GAIN, TAG_REPLAYGAIN_ALBUM_PEAK, TAG_REPLAYGAIN_TRACK_GAIN,
-    TAG_REPLAYGAIN_TRACK_PEAK,
+    Channel, TAG_MP3GAIN_MINMAX, TAG_MP3GAIN_UNDO, TAG_REPLAYGAIN_ALBUM_GAIN,
+    TAG_REPLAYGAIN_ALBUM_PEAK, TAG_REPLAYGAIN_TRACK_GAIN, TAG_REPLAYGAIN_TRACK_PEAK,
 };
 use std::env;
-use std::fs;
 use std::path::{Path, PathBuf};
-use std::time::SystemTime;
 
 use cli::options::{AacAlbumInfo, Options, OutputFormat, StoredTagMode};
 use cli::parse_args::{expand_files_recursive, parse_args};
 use cli::usage::print_usage;
 use json_output::{JsonAlbumResult, JsonFileResult, JsonOutput, JsonSummary};
+use processors::apply::{process_apply, process_apply_channel};
+use processors::info::process_info;
+use processors::replaygain::{process_apply_replaygain_with_album, process_track_gain};
+use processors::undo::process_undo;
+use processors::utils::restore_timestamp;
 use progress::{
     create_analysis_progress_bar, create_progress_bar, finish_analysis_progress, progress_finish,
-    progress_inc, progress_set_message, update_analysis_progress, PROGRESS_THRESHOLD,
+    progress_inc, progress_set_message, PROGRESS_THRESHOLD,
 };
 
 /// Extract filename from path, returning "unknown" if extraction fails
-fn get_filename(path: &Path) -> &str {
+pub(crate) fn get_filename(path: &Path) -> &str {
     path.file_name()
         .and_then(|n| n.to_str())
         .unwrap_or("unknown")
@@ -1254,1054 +1256,4 @@ fn cmd_album_gain(files: &[PathBuf], opts: &Options) -> Result<()> {
     }
 
     Ok(())
-}
-
-// =============================================================================
-// File processing
-// =============================================================================
-
-fn apply_with_temp_file<F>(file: &Path, operation: F, opts: &Options) -> Result<usize>
-where
-    F: FnOnce(&Path) -> Result<usize>,
-{
-    if opts.use_temp_file {
-        // Create temp file in the same directory
-        let parent = file.parent().unwrap_or(Path::new("."));
-        let temp_path = parent.join(format!(".mp3rgain_temp_{}.mp3", std::process::id()));
-
-        // Copy original to temp
-        fs::copy(file, &temp_path)?;
-
-        // Apply operation to temp file
-        match operation(&temp_path) {
-            Ok(frames) => {
-                // Replace original with temp
-                fs::rename(&temp_path, file)?;
-                Ok(frames)
-            }
-            Err(e) => {
-                // Clean up temp file on error
-                let _ = fs::remove_file(&temp_path);
-                Err(e)
-            }
-        }
-    } else {
-        operation(file)
-    }
-}
-
-fn process_apply(file: &Path, steps: i32, opts: &Options) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-    let dry_run_prefix = opts.dry_run_prefix();
-
-    // Save original timestamp if needed
-    let original_mtime = if opts.preserve_timestamp && !opts.dry_run {
-        std::fs::metadata(file).ok().and_then(|m| m.modified().ok())
-    } else {
-        None
-    };
-
-    let is_aac = mp4meta::is_aac_file(file);
-
-    // Multi-track warning for AAC files
-    if is_aac && opts.output_format == OutputFormat::Text && !opts.quiet {
-        let track_count = mp4meta::count_audio_tracks(file);
-        if track_count > 1 {
-            eprintln!(
-                "  {} {}{} - {} audio tracks detected, processing first track only",
-                "!".yellow(),
-                dry_run_prefix,
-                filename,
-                track_count
-            );
-        }
-    }
-
-    // Check for clipping and possibly prevent it
-    let mut actual_steps = steps;
-    let mut warning_msg: Option<String> = None;
-
-    if steps > 0 && !opts.wrap_gain {
-        let headroom = if is_aac {
-            aac::analyze_aac_gains(file)
-                .ok()
-                .map(|a| (255u8.saturating_sub(a.max_gain())) as i32)
-        } else {
-            analyze(file).ok().map(|info| info.headroom_steps())
-        };
-
-        if let Some(headroom_steps) = headroom {
-            if steps > headroom_steps {
-                if opts.prevent_clipping {
-                    let original_steps = steps;
-                    actual_steps = headroom_steps;
-                    if opts.output_format == OutputFormat::Text && !opts.quiet {
-                        eprintln!(
-                            "  {} {}{} - gain reduced from {} to {} steps to prevent clipping",
-                            "!".yellow(),
-                            dry_run_prefix,
-                            filename,
-                            original_steps,
-                            actual_steps
-                        );
-                    }
-                    warning_msg = Some(format!(
-                        "gain reduced from {} to {} steps to prevent clipping",
-                        original_steps, actual_steps
-                    ));
-                } else if !opts.ignore_clipping && !opts.quiet {
-                    if opts.output_format == OutputFormat::Text {
-                        eprintln!(
-                            "  {} {}{} - clipping warning: requested {} steps but only {} headroom",
-                            "!".yellow(),
-                            dry_run_prefix,
-                            filename,
-                            steps,
-                            headroom_steps
-                        );
-                        eprintln!(
-                            "      Use -c to ignore clipping warnings or -k to prevent clipping"
-                        );
-                    }
-                    warning_msg = Some(format!(
-                        "clipping warning: requested {} steps but only {} headroom",
-                        steps, headroom_steps
-                    ));
-                }
-            }
-        }
-    }
-
-    // Dry run: don't actually modify
-    if opts.dry_run {
-        if opts.output_format == OutputFormat::Text && !opts.quiet {
-            println!(
-                "  {} [DRY RUN] {} (would apply {} steps)",
-                "~".cyan(),
-                filename,
-                actual_steps
-            );
-        }
-        return Ok(JsonFileResult {
-            file: file.display().to_string(),
-            status: Some("dry_run".to_string()),
-            gain_applied_steps: Some(actual_steps),
-            gain_applied_db: Some(steps_to_db(actual_steps)),
-            warning: warning_msg,
-            dry_run: Some(true),
-            ..Default::default()
-        });
-    }
-
-    // Apply gain
-    let apply_result = if is_aac {
-        if opts.stored_tag_mode == StoredTagMode::Skip {
-            apply_with_temp_file(file, |f| Ok(aac::apply_aac_gain(f, actual_steps)?), opts)
-        } else {
-            apply_with_temp_file(
-                file,
-                |f| Ok(aac::apply_aac_gain_with_undo(f, actual_steps)?),
-                opts,
-            )
-        }
-    } else if opts.use_id3v2 {
-        // MP3 with -s i: apply gain without APE undo, write undo to ID3v2
-        let skip_undo = opts.stored_tag_mode == StoredTagMode::Skip;
-        let result = apply_with_temp_file(
-            file,
-            |f| {
-                Ok(GainOptions::new(actual_steps)
-                    .wrap(opts.wrap_gain)
-                    .undo(false)
-                    .apply(f)?)
-            },
-            opts,
-        );
-        if !skip_undo && result.is_ok() {
-            let analysis = analyze(file)?;
-            let existing_rg = id3v2::read_id3v2_replaygain(file).unwrap_or_default();
-            let (existing_left, _) = mp3rgain::ape::parse_undo_values(existing_rg.undo.as_deref());
-            let new_undo = existing_left + actual_steps;
-            id3v2::write_id3v2_undo(
-                file,
-                new_undo,
-                new_undo,
-                opts.wrap_gain,
-                analysis.min_gain(),
-                analysis.max_gain(),
-            )?;
-        }
-        result
-    } else {
-        let use_undo = opts.stored_tag_mode != StoredTagMode::Skip;
-        apply_with_temp_file(
-            file,
-            |f| {
-                Ok(GainOptions::new(actual_steps)
-                    .wrap(opts.wrap_gain)
-                    .undo(use_undo)
-                    .apply(f)?)
-            },
-            opts,
-        )
-    };
-
-    match apply_result {
-        Ok(modified) => {
-            // Restore timestamp if needed
-            if let Some(mtime) = original_mtime {
-                restore_timestamp(file, mtime);
-            }
-
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                if is_aac {
-                    println!(
-                        "  {} {} ({} gains modified)",
-                        "v".green(),
-                        filename,
-                        modified
-                    );
-                } else {
-                    println!("  {} {} ({} frames)", "v".green(), filename, modified);
-                }
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("success".to_string()),
-                frames: Some(modified),
-                gain_applied_steps: Some(actual_steps),
-                gain_applied_db: Some(steps_to_db(actual_steps)),
-                warning: warning_msg,
-                ..Default::default()
-            })
-        }
-        Err(e) => {
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                eprintln!("  {} {} - {}", "x".red(), filename, e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-fn process_apply_channel(
-    file: &Path,
-    channel: Channel,
-    steps: i32,
-    opts: &Options,
-) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-    let channel_name = match channel {
-        Channel::Left => "left",
-        Channel::Right => "right",
-        _ => unreachable!(),
-    };
-
-    // Warn if file is Joint Stereo (mp3gain only supports Stereo for -l)
-    if let Ok(info) = analyze(file) {
-        if info.channel_mode() == mp3rgain::ChannelMode::JointStereo
-            && opts.output_format == OutputFormat::Text
-            && !opts.quiet
-        {
-            eprintln!(
-                "  {} {} - Joint Stereo file: channel-specific gain may not work as expected",
-                "!".yellow(),
-                filename
-            );
-        }
-    }
-
-    // Save original timestamp if needed
-    let original_mtime = if opts.preserve_timestamp && !opts.dry_run {
-        std::fs::metadata(file).ok().and_then(|m| m.modified().ok())
-    } else {
-        None
-    };
-
-    // Dry run: don't actually modify
-    if opts.dry_run {
-        if opts.output_format == OutputFormat::Text && !opts.quiet {
-            println!(
-                "  {} [DRY RUN] {} (would apply {} steps to {} channel)",
-                "~".cyan(),
-                filename,
-                steps,
-                channel_name
-            );
-        }
-        return Ok(JsonFileResult {
-            file: file.display().to_string(),
-            status: Some("dry_run".to_string()),
-            gain_applied_steps: Some(steps),
-            gain_applied_db: Some(steps_to_db(steps)),
-            dry_run: Some(true),
-            ..Default::default()
-        });
-    }
-
-    let apply_result = if opts.use_id3v2 {
-        // Apply gain without APE undo, then write undo to ID3v2
-        let result = GainOptions::new(steps)
-            .channel(channel)
-            .undo(false)
-            .apply(file);
-        if result.is_ok() {
-            let analysis = analyze(file)?;
-            let existing_rg = id3v2::read_id3v2_replaygain(file).unwrap_or_default();
-            let (existing_left, existing_right) =
-                mp3rgain::ape::parse_undo_values(existing_rg.undo.as_deref());
-            let (new_left, new_right) = match channel {
-                Channel::Left => (existing_left + steps, existing_right),
-                Channel::Right => (existing_left, existing_right + steps),
-                _ => unreachable!(),
-            };
-            id3v2::write_id3v2_undo(
-                file,
-                new_left,
-                new_right,
-                false,
-                analysis.min_gain(),
-                analysis.max_gain(),
-            )?;
-        }
-        result
-    } else {
-        GainOptions::new(steps)
-            .channel(channel)
-            .undo(true)
-            .apply(file)
-    };
-
-    match apply_result {
-        Ok(frames) => {
-            // Restore timestamp if needed
-            if let Some(mtime) = original_mtime {
-                restore_timestamp(file, mtime);
-            }
-
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                println!(
-                    "  {} {} ({} frames, {} channel)",
-                    "v".green(),
-                    filename,
-                    frames,
-                    channel_name
-                );
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("success".to_string()),
-                frames: Some(frames),
-                gain_applied_steps: Some(steps),
-                gain_applied_db: Some(steps_to_db(steps)),
-                ..Default::default()
-            })
-        }
-        Err(e) => {
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                eprintln!("  {} {} - {}", "x".red(), filename, e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-fn process_info(
-    file: &Path,
-    opts: &Options,
-    analysis_pb: Option<&ProgressBar>,
-) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-
-    // Perform ReplayGain analysis for TSV/Text output (mp3gain compatible)
-    if matches!(opts.output_format, OutputFormat::Tsv | OutputFormat::Text)
-        && replaygain::is_available()
-    {
-        let rg_result = if let Some(pb) = analysis_pb {
-            replaygain::analyze_track_with_progress(file, opts.track_index, &|bytes, total| {
-                update_analysis_progress(&Some(pb.clone()), bytes, total);
-            })
-        } else {
-            replaygain::analyze_track_with_index(file, opts.track_index)
-        };
-
-        match rg_result {
-            Ok(rg_result) => {
-                // Get max amplitude info
-                let (max_amp, max_gain, min_gain) = find_max_amplitude(file)
-                    .map(|r| (r.max_amplitude(), r.max_global_gain(), r.min_global_gain()))
-                    .unwrap_or((1.0, 255, 0));
-
-                // Calculate gain with modifier (mp3gain compatible: -d modifies suggested gain)
-                let gain_db = rg_result.gain_db() + opts.gain_modifier_db;
-                let gain_steps = db_to_steps(gain_db);
-
-                // Max Amplitude scaled to 32768 (mp3gain format for beets)
-                // beets divides by 32768, so we output peak * 32768
-                let max_amplitude_scaled = rg_result.peak() * 32768.0;
-
-                match opts.output_format {
-                    OutputFormat::Tsv => {
-                        println!(
-                            "{}\t{}\t{:.6}\t{:.6}\t{}\t{}",
-                            filename, gain_steps, gain_db, max_amplitude_scaled, max_gain, min_gain
-                        );
-                    }
-                    OutputFormat::Text => {
-                        if !opts.quiet {
-                            println!("{}", filename.cyan().bold());
-                            println!("  Recommended \"Track\" dB change: {:.6}", gain_db);
-                            println!("  Recommended \"Track\" mp3 gain change: {}", gain_steps);
-                            println!(
-                                "  Max PCM sample at current gain: {:.6}",
-                                max_amplitude_scaled
-                            );
-                            println!("  Max mp3 global gain field: {}", max_gain);
-                            println!("  Min mp3 global gain field: {}", min_gain);
-                            println!();
-                        }
-                    }
-                    OutputFormat::Json => {}
-                }
-
-                return Ok(JsonFileResult {
-                    file: file.display().to_string(),
-                    loudness_db: Some(rg_result.loudness_db()),
-                    gain_applied_db: Some(gain_db),
-                    gain_applied_steps: Some(gain_steps),
-                    peak: Some(rg_result.peak()),
-                    max_amplitude: Some(max_amp),
-                    max_gain: Some(max_gain),
-                    min_gain: Some(min_gain),
-                    ..Default::default()
-                });
-            }
-            Err(e) => {
-                eprintln!("{} - {}", filename.red(), e);
-                return Ok(JsonFileResult {
-                    file: file.display().to_string(),
-                    status: Some("error".to_string()),
-                    error: Some(e.to_string()),
-                    ..Default::default()
-                });
-            }
-        }
-    }
-
-    // Check if this is an M4A/AAC file - if so, show appropriate message
-    if mp4meta::is_mp4_file(file) {
-        let codec = mp4meta::detect_mp4_audio_codec(file);
-        let is_alac = matches!(codec, Some(mp4meta::Mp4AudioCodec::Alac));
-        let format_str = if is_alac { "M4A/ALAC" } else { "M4A/AAC" };
-
-        match opts.output_format {
-            OutputFormat::Text => {
-                if opts.quiet {
-                    println!("{}\t{}\t-\t-\t-\t-\t-", filename, format_str);
-                } else {
-                    println!("{}", filename.cyan().bold());
-                    println!("  Format:      {}", format_str);
-                    if is_alac {
-                        println!(
-                            "  {}",
-                            "ALAC files are not supported for gain adjustment".yellow()
-                        );
-                    } else {
-                        println!(
-                            "  {}",
-                            "Note: Use -r or -a for ReplayGain analysis".yellow()
-                        );
-                    }
-                    println!();
-                }
-            }
-            OutputFormat::Tsv => {
-                println!("{}\t-\t-\t-\t-\t-", filename);
-            }
-            OutputFormat::Json => {}
-        }
-
-        return Ok(JsonFileResult {
-            file: file.display().to_string(),
-            status: Some("info".to_string()),
-            ..Default::default()
-        });
-    }
-
-    // MP3 file: use basic analysis
-    match analyze(file) {
-        Ok(info) => {
-            match opts.output_format {
-                OutputFormat::Text => {
-                    if opts.quiet {
-                        // Quiet mode: tab-separated output
-                        println!(
-                            "{}\t{}\t{}\t{}\t{:.1}\t{}\t{:.1}",
-                            filename,
-                            info.frame_count(),
-                            info.min_gain(),
-                            info.max_gain(),
-                            info.avg_gain(),
-                            info.headroom_steps(),
-                            info.headroom_db()
-                        );
-                    } else {
-                        println!("{}", filename.cyan().bold());
-                        println!(
-                            "  Format:      {} Layer III, {}",
-                            info.mpeg_version(),
-                            info.channel_mode()
-                        );
-                        println!("  Frames:      {}", info.frame_count());
-                        println!(
-                            "  Gain range:  {} - {} (avg: {:.1})",
-                            info.min_gain(),
-                            info.max_gain(),
-                            info.avg_gain()
-                        );
-                        println!(
-                            "  Headroom:    {} steps ({:+.1} dB)",
-                            info.headroom_steps().to_string().green(),
-                            info.headroom_db()
-                        );
-                        println!();
-                    }
-                }
-                OutputFormat::Tsv => {
-                    // Fallback TSV (ReplayGain not available): basic info
-                    println!(
-                        "{}\t{}\t{:.1}\t{:.6}\t{}\t{}",
-                        filename,
-                        info.headroom_steps(),
-                        info.headroom_db(),
-                        1.0,
-                        info.max_gain(),
-                        info.min_gain()
-                    );
-                }
-                OutputFormat::Json => {}
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                mpeg_version: Some(info.mpeg_version().to_string()),
-                channel_mode: Some(info.channel_mode().to_string()),
-                frames: Some(info.frame_count()),
-                min_gain: Some(info.min_gain()),
-                max_gain: Some(info.max_gain()),
-                avg_gain: Some(info.avg_gain()),
-                headroom_steps: Some(info.headroom_steps()),
-                headroom_db: Some(info.headroom_db()),
-                ..Default::default()
-            })
-        }
-        Err(e) => {
-            if opts.output_format != OutputFormat::Json {
-                eprintln!("{} - {}", filename.red(), e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-fn process_undo(file: &Path, opts: &Options) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-    let dry_run_prefix = opts.dry_run_prefix();
-
-    // Save original timestamp if needed
-    let original_mtime = if opts.preserve_timestamp && !opts.dry_run {
-        std::fs::metadata(file).ok().and_then(|m| m.modified().ok())
-    } else {
-        None
-    };
-
-    // Dry run: just analyze what would be done
-    if opts.dry_run {
-        // Try to read the undo tag to see what would happen
-        if opts.output_format == OutputFormat::Text && !opts.quiet {
-            println!("  {} [DRY RUN] {} (would undo)", "~".cyan(), filename);
-        }
-        return Ok(JsonFileResult {
-            file: file.display().to_string(),
-            status: Some("dry_run".to_string()),
-            dry_run: Some(true),
-            ..Default::default()
-        });
-    }
-
-    let is_aac = mp4meta::is_aac_file(file);
-    let undo_result = if is_aac {
-        aac::undo_aac_gain(file)
-    } else if opts.use_id3v2 {
-        id3v2::undo_gain_id3v2(file)
-    } else {
-        undo_gain(file)
-    };
-
-    match undo_result {
-        Ok(frames) => {
-            if frames == 0 {
-                if opts.output_format == OutputFormat::Text && !opts.quiet {
-                    println!(
-                        "  {} {}{} (no changes to undo)",
-                        ".".cyan(),
-                        dry_run_prefix,
-                        filename
-                    );
-                }
-
-                Ok(JsonFileResult {
-                    file: file.display().to_string(),
-                    status: Some("skipped".to_string()),
-                    frames: Some(0),
-                    ..Default::default()
-                })
-            } else {
-                // Restore timestamp if needed
-                if let Some(mtime) = original_mtime {
-                    restore_timestamp(file, mtime);
-                }
-
-                if opts.output_format == OutputFormat::Text && !opts.quiet {
-                    println!(
-                        "  {} {} ({} frames restored)",
-                        "v".green(),
-                        filename,
-                        frames
-                    );
-                }
-
-                Ok(JsonFileResult {
-                    file: file.display().to_string(),
-                    status: Some("success".to_string()),
-                    frames: Some(frames),
-                    ..Default::default()
-                })
-            }
-        }
-        Err(e) => {
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                eprintln!("  {} {} - {}", "x".red(), filename, e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-fn process_track_gain(
-    file: &Path,
-    opts: &Options,
-    analysis_pb: Option<&ProgressBar>,
-) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-    let dry_run_prefix = opts.dry_run_prefix();
-
-    if opts.output_format == OutputFormat::Text && !opts.quiet {
-        println!(
-            "  {} {}Analyzing {}...",
-            "->".cyan(),
-            dry_run_prefix,
-            filename
-        );
-    }
-
-    let rg_result = if let Some(pb) = analysis_pb {
-        replaygain::analyze_track_with_progress(file, opts.track_index, &|bytes, total| {
-            update_analysis_progress(&Some(pb.clone()), bytes, total);
-        })
-    } else {
-        replaygain::analyze_track_with_index(file, opts.track_index)
-    };
-
-    match rg_result {
-        Ok(result) => {
-            // Apply gain modifier
-            let base_steps = result.gain_steps();
-            let modified_steps = base_steps + opts.gain_modifier;
-
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                println!(
-                    "      Loudness: {:.1} dB, Gain: {:+.1} dB ({} steps{}), Peak: {:.4}",
-                    result.loudness_db(),
-                    result.gain_db(),
-                    base_steps,
-                    if opts.gain_modifier != 0 {
-                        format!(" + {} = {}", opts.gain_modifier, modified_steps)
-                    } else {
-                        String::new()
-                    },
-                    result.peak()
-                );
-            }
-
-            if modified_steps == 0 {
-                if opts.output_format == OutputFormat::Text && !opts.quiet {
-                    println!("  {} {} (no adjustment needed)", ".".cyan(), filename);
-                }
-                return Ok(JsonFileResult {
-                    file: file.display().to_string(),
-                    status: Some("skipped".to_string()),
-                    loudness_db: Some(result.loudness_db()),
-                    peak: Some(result.peak()),
-                    gain_applied_steps: Some(0),
-                    gain_applied_db: Some(0.0),
-                    ..Default::default()
-                });
-            }
-
-            process_apply_replaygain_with_album(file, modified_steps, &result, opts, None)
-        }
-        Err(e) => {
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                eprintln!("  {} {} - {}", "x".red(), filename, e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-fn process_apply_replaygain_with_album(
-    file: &Path,
-    steps: i32,
-    result: &ReplayGainResult,
-    opts: &Options,
-    album_info: Option<&AacAlbumInfo>,
-) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-    let dry_run_prefix = opts.dry_run_prefix();
-
-    // Save original timestamp if needed
-    let original_mtime = if opts.preserve_timestamp && !opts.dry_run {
-        std::fs::metadata(file).ok().and_then(|m| m.modified().ok())
-    } else {
-        None
-    };
-
-    // Check for clipping if not ignored
-    let mut actual_steps = steps;
-    let mut warning_msg: Option<String> = None;
-
-    if steps > 0 && !opts.wrap_gain {
-        // Check if applying this gain would cause clipping
-        let gain_linear = 10.0_f64.powf(result.gain_db() / 20.0);
-        let new_peak = result.peak() * gain_linear;
-        if new_peak > 1.0 {
-            if opts.prevent_clipping {
-                // Calculate the maximum safe gain
-                let max_safe_db = -20.0 * result.peak().log10();
-                let max_safe_steps = db_to_steps(max_safe_db);
-                actual_steps = max_safe_steps.max(0);
-
-                if opts.output_format == OutputFormat::Text && !opts.quiet {
-                    eprintln!(
-                        "  {} {}{} - gain reduced from {} to {} steps to prevent clipping (peak: {:.4})",
-                        "!".yellow(),
-                        dry_run_prefix,
-                        filename,
-                        steps,
-                        actual_steps,
-                        result.peak()
-                    );
-                }
-                warning_msg = Some(format!(
-                    "gain reduced from {} to {} steps to prevent clipping (peak: {:.4})",
-                    steps,
-                    actual_steps,
-                    result.peak()
-                ));
-            } else if !opts.ignore_clipping && !opts.quiet {
-                if opts.output_format == OutputFormat::Text {
-                    eprintln!(
-                        "  {} {}{} - clipping warning: peak would be {:.2} (>{:.2})",
-                        "!".yellow(),
-                        dry_run_prefix,
-                        filename,
-                        new_peak,
-                        1.0
-                    );
-                    eprintln!("      Use -c to ignore clipping warnings or -k to prevent clipping");
-                }
-                warning_msg = Some(format!(
-                    "clipping warning: peak would be {:.2} (>1.00)",
-                    new_peak
-                ));
-            }
-        }
-    }
-
-    // Dry run: don't actually modify
-    if opts.dry_run {
-        if opts.output_format == OutputFormat::Text && !opts.quiet {
-            println!(
-                "  {} [DRY RUN] {} (would apply {:+.1} dB, {} steps)",
-                "~".cyan(),
-                filename,
-                steps_to_db(actual_steps),
-                actual_steps,
-            );
-        }
-        return Ok(JsonFileResult {
-            file: file.display().to_string(),
-            status: Some("dry_run".to_string()),
-            loudness_db: Some(result.loudness_db()),
-            peak: Some(result.peak()),
-            gain_applied_steps: Some(actual_steps),
-            gain_applied_db: Some(steps_to_db(actual_steps)),
-            warning: warning_msg,
-            dry_run: Some(true),
-            ..Default::default()
-        });
-    }
-
-    // Handle AAC/M4A files differently - only write ReplayGain tags
-    if result.file_type() == AudioFileType::Aac {
-        return process_apply_replaygain_aac_with_album(
-            file,
-            actual_steps,
-            result,
-            opts,
-            warning_msg,
-            original_mtime,
-            album_info,
-        );
-    }
-
-    // MP3: Apply gain to audio frames
-    let use_id3v2_undo = opts.use_id3v2;
-    let apply_result = apply_with_temp_file(
-        file,
-        |f| {
-            Ok(GainOptions::new(actual_steps)
-                .wrap(opts.wrap_gain)
-                .undo(!use_id3v2_undo) // APE undo only when not using ID3v2
-                .apply(f)?)
-        },
-        opts,
-    );
-
-    match apply_result {
-        Ok(frames) => {
-            // Write ID3v2 tags if -s i mode
-            if opts.use_id3v2 {
-                let analysis = analyze(file)?;
-                let existing_rg = id3v2::read_id3v2_replaygain(file).unwrap_or_default();
-                let (existing_left, _) =
-                    mp3rgain::ape::parse_undo_values(existing_rg.undo.as_deref());
-                let new_undo = existing_left + actual_steps;
-
-                // Write undo data
-                id3v2::write_id3v2_undo(
-                    file,
-                    new_undo,
-                    new_undo,
-                    opts.wrap_gain,
-                    analysis.min_gain(),
-                    analysis.max_gain(),
-                )?;
-
-                // Write ReplayGain metadata tags
-                let rg = mp3rgain::Id3v2ReplayGain {
-                    track_gain: Some(format!("{:+.2} dB", result.gain_db())),
-                    track_peak: Some(format!("{:.6}", result.peak())),
-                    album_gain: album_info.map(|a| format!("{:+.2} dB", a.album_gain_db)),
-                    album_peak: album_info.map(|a| format!("{:.6}", a.album_peak)),
-                    ..Default::default()
-                };
-                id3v2::write_id3v2_replaygain(file, &rg)?;
-            }
-
-            // Restore timestamp if needed
-            if let Some(mtime) = original_mtime {
-                restore_timestamp(file, mtime);
-            }
-
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                println!(
-                    "  {} {} ({} frames, {:+.1} dB)",
-                    "v".green(),
-                    filename,
-                    frames,
-                    steps_to_db(actual_steps)
-                );
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("success".to_string()),
-                frames: Some(frames),
-                loudness_db: Some(result.loudness_db()),
-                peak: Some(result.peak()),
-                gain_applied_steps: Some(actual_steps),
-                gain_applied_db: Some(steps_to_db(actual_steps)),
-                warning: warning_msg,
-                ..Default::default()
-            })
-        }
-        Err(e) => {
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                eprintln!("  {} {} - {}", "x".red(), filename, e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-/// Apply ReplayGain to AAC/M4A files with optional album info
-fn process_apply_replaygain_aac_with_album(
-    file: &Path,
-    actual_steps: i32,
-    result: &ReplayGainResult,
-    opts: &Options,
-    warning_msg: Option<String>,
-    original_mtime: Option<std::time::SystemTime>,
-    album_info: Option<&AacAlbumInfo>,
-) -> Result<JsonFileResult> {
-    let filename = get_filename(file);
-
-    // Multi-track warning
-    if opts.output_format == OutputFormat::Text && !opts.quiet {
-        let track_count = mp4meta::count_audio_tracks(file);
-        if track_count > 1 {
-            eprintln!(
-                "  {} {} - {} audio tracks detected, processing first track only",
-                "!".yellow(),
-                filename,
-                track_count
-            );
-        }
-    }
-
-    // Apply bitstream gain modification with undo support
-    let gain_modified = if actual_steps != 0 {
-        match aac::apply_aac_gain_with_undo(file, actual_steps) {
-            Ok(n) => n,
-            Err(e) => {
-                if opts.output_format == OutputFormat::Text && !opts.quiet {
-                    eprintln!(
-                        "  {} {} - bitstream gain failed: {} (tags still written)",
-                        "!".yellow(),
-                        filename,
-                        e
-                    );
-                }
-                0
-            }
-        }
-    } else {
-        0
-    };
-
-    // Create ReplayGain tags for AAC
-    let mut tags = mp4meta::ReplayGainTags::default();
-    tags.set_track(result.gain_db(), result.peak());
-
-    // Add album tags if available
-    if let Some(album) = album_info {
-        tags.set_album(album.album_gain_db, album.album_peak);
-    }
-
-    // Write tags to file
-    match mp4meta::write_replaygain_tags(file, &tags) {
-        Ok(()) => {
-            // Restore timestamp if needed
-            if let Some(mtime) = original_mtime {
-                restore_timestamp(file, mtime);
-            }
-
-            let tag_type = if album_info.is_some() {
-                "track+album tags"
-            } else {
-                "tags"
-            };
-
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                if gain_modified > 0 {
-                    println!(
-                        "  {} {} ({} gains modified + {} written, {:+.1} dB)",
-                        "v".green(),
-                        filename,
-                        gain_modified,
-                        tag_type,
-                        result.gain_db()
-                    );
-                } else {
-                    println!(
-                        "  {} {} ({} written, {:+.1} dB)",
-                        "v".green(),
-                        filename,
-                        tag_type,
-                        result.gain_db()
-                    );
-                }
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("success".to_string()),
-                loudness_db: Some(result.loudness_db()),
-                peak: Some(result.peak()),
-                gain_applied_steps: Some(actual_steps),
-                gain_applied_db: Some(steps_to_db(actual_steps)),
-                warning: warning_msg,
-                ..Default::default()
-            })
-        }
-        Err(e) => {
-            if opts.output_format == OutputFormat::Text && !opts.quiet {
-                eprintln!("  {} {} - {}", "x".red(), filename, e);
-            }
-
-            Ok(JsonFileResult {
-                file: file.display().to_string(),
-                status: Some("error".to_string()),
-                error: Some(e.to_string()),
-                ..Default::default()
-            })
-        }
-    }
-}
-
-fn restore_timestamp(file: &Path, mtime: SystemTime) {
-    let _ = std::fs::File::options()
-        .write(true)
-        .open(file)
-        .and_then(|f| f.set_times(std::fs::FileTimes::new().set_modified(mtime)));
 }
