+++ src/processors/mod.rs
@@ -1,0 +1,5 @@
+pub mod apply;
+pub mod info;
+pub mod replaygain;
+pub mod undo;
+pub mod utils;
